requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
httpx==0.28.1
feedparser==6.0.11
openai==1.59.9
google-generativeai==0.8.4
python-multipart==0.0.20
sumy==0.11.0
numpy==1.26.4
nltk==3.9.1
anthropic==0.42.0
